+++ RELEASE-NOTES.md
@@ -1,9 +1,9 @@
 V3.0.0
 ======
-- Standarize the connectionName in the UI and backend  ???
-- Abstract colors and various Atoms  
-- Create the predefined queries store and APIs  
+- Create the predefined queries store and APIs [V see truncate table/empty table operation as a working example]
+
 - Create a standalone app launch on top of php -S
 - Enable deep links
 - table operation page  
 - Paged query results  
+- Abstract colors and various Atoms